requirements.txt
qbittorrent-api
anitopy
psutil
filelock
tmdbv3api
python-plexapi
